requirements.txt
datasets==2.18.0
dkpro_cassis==0.9.1
langdetect==1.0.9
matplotlib==3.9.3
networkx==3.3
nltk==3.8.1
numpy==2.1.3
openai==1.55.3
pandas==2.2.3
scikit_learn==1.5.2
scipy==1.14.1
sentence_transformers==3.0.1
spacy==3.7.2
tqdm==4.66.1
